test2/requirement.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pyrsistent==0.19.3
pytest==7.4.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.17
typing_extensions==4.7.0
Werkzeug==2.3.6
